#1341 - Move Builder Component classes into separate folder
