+++ src/Builder/Exception/BuilderException.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace Phalcon\DevTools\Builder;
+namespace Phalcon\DevTools\Builder\Exception;
 
 use Phalcon\Exception;
