+++ src/Web/Tools/Controllers/MigrationsController.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 namespace Phalcon\DevTools\Web\Tools\Controllers;
 
 use DirectoryIterator;
-use Phalcon\DevTools\Builder\BuilderException;
+use Phalcon\DevTools\Builder\Exception\BuilderException;
 use Phalcon\DevTools\Mvc\Controller\Base;
 use Phalcon\Flash\Direct;
 use Phalcon\Flash\Session;
